Finished initial implementation of help system. It consists of the following:
1. A web service (web/help) that exports the help menu configuration. The help menu
information is built from a layout.xml file (in core-help) that determines the
structure of the menu and from many content.xml files in modules that populate the
menu content. The web service also has a deployer that listens for module parts
of type "help".

2. A core-help module (modules/world/core-help) that defines a test structure for
the menu system. This is where "core" help will go.

3. Enhancements to the WL client to read the help menu information from the
web server and create the JMenu and launch an external browser.