Text chat support for Wonderland. Supports both "global" chat (everyone) and person-to-person text chat. Added a "Text Chat" menu under "Tools" that has two checkbox items, to display the "global" chat window and a window to display a list of users to launch chat sessions.
Added new connection type, handled on the server with TextChatConnectionHandler and registered via a server
plugin. The client connection is represented by TextChatConnection.

There is a single message: TextChatMessage that contains the String message, and the from/to. (For a "global"
message, the "to" is an empty string).

Registers a component on avatars for a context menu to launch a chat with that user. This does not yet work,
since it seems the context menu does not get events for avatars.