Fix Issue #523: Provide a dialog box that allows users to select among multiple Cells (err... Components in the GUI) that support a given extension (e.g. PDF Viewer and PDF Slider Spreader) to use.
- Added CellSelectionRegistry.java and CellSelectionSPI.java that provides a registry and SPI interface so modules can
  register to determine how to resolve which Cell to use. CellSelectionRegistry only supports a single CellSelectionSPI
  object.
- Added the cell-selection module that implements CellSelectionSPI and displays a simple dialog box to allow users to
  select among the choices. (This currently does not remember the choices made, left for future work that would also
  need to include a way to edit the extension -> Cell mappings and store in WebDav).

An important note: The cell-selection dialog box uses CellFactorySPI.getDisplayName() for the string to display in its
list. However, Cell factories may return null for this method so they do not appear in the Cell Palettes, but they
would still need to appear in the cell-selection dialog box list. The correct solution, which cannot happen for 0.5
because it is an API change with fairly wide impact is to augment CellFactorySPI. For now, the cell-selection dialog box
just uses the result of toString() from the Cell factory class. (Overrided toString in ImageViewerCellFactory so
that it returns something sensible).