Improved softphone error reporting.
Changes for VuMeter.

Make sendMessage() public instead of protected in VoiceBridgeConnection.

Added listeners to get voice bridge audio and rtp data.

Added a flag to AudioGroupSetup to not remove the group automatically
when the last member leaves the group.

Print a warning message if the voice bridge public address is unusable.